Refactoring ContrabandHitRateBar
